fix: #293 refactoring later
